Refactor trainer package to trainerdetail. Remove unused viewmodel.
