+++ app/src/main/res/navigation/nav_main.xml
@@ -128,7 +128,7 @@
 
     <fragment
         android:id="@+id/trainer_fragment"
-        android:name="com.trx.consumer.screens.trainer.TrainerDetailFragment"
+        android:name="com.trx.consumer.screens.trainerdetail.TrainerDetailFragment"
         android:label="TrainerFragment"
         tools:layout="@layout/fragment_trainer_detail" />
